+++ src/components/Input/InputContainer.tsx
@@ -3,12 +3,14 @@ import Input from "./Input";
 import AddButton from "../Button/AddButton";
 import { Dispatch, FormEvent, useState } from "react";
 import { TodoProps } from "../../types/TodoProps";
+import { LocalStorageManager } from "../../util/localStorageManager";
 
-const InputContainer = ({
-  setInitTodo,
-}: {
+interface InputContainerProps {
   setInitTodo: Dispatch<React.SetStateAction<TodoProps[]>>;
-}) => {
+  addTodo: (inputData: string, storage: LocalStorageManager) => void;
+}
+
+const InputContainer = ({ setInitTodo, addTodo }: InputContainerProps) => {
   const [inputData, setInputData] = useState("");
 
   const onSubmit = (e: FormEvent<HTMLFormElement>) => {
@@ -19,6 +21,7 @@ const InputContainer = ({
     <InputContainerComponent onSubmit={(e) => onSubmit(e)}>
       <Input setInputData={setInputData} inputData={inputData} />
       <AddButton
+        addTodo={addTodo}
         inputData={inputData}
         setInitTodo={setInitTodo}
         setInputData={setInputData}
